refactor(backend): share readS3ObjectAsBuffer + harden theme read [GLITCH-569]
