+++ packages/backend/src/ee/services/AppGenerateService/readDesignForDownload.ts
@@ -1,34 +1,10 @@
-import { GetObjectCommand, S3Client } from '@aws-sdk/client-s3';
+import { type S3Client } from '@aws-sdk/client-s3';
 import { type DataAppThemeContext } from '@lightdash/common';
 import type { Logger } from 'winston';
 import type { OrganizationDesignModel } from '../../../models/OrganizationDesignModel';
 import { designS3Key } from '../../../services/OrganizationDesignService/OrganizationDesignService';
 import { contextFile, THEME_ASSET_CAP } from './appContext';
-
-const readS3ObjectAsBuffer = async (
-    s3Client: S3Client,
-    bucket: string,
-    key: string,
-): Promise<Buffer> => {
-    const response = await s3Client.send(
-        new GetObjectCommand({ Bucket: bucket, Key: key }),
-    );
-    const body = response.Body;
-    if (!body || typeof (body as NodeJS.ReadableStream).on !== 'function') {
-        throw new Error(`Unexpected S3 response body type for key=${key}`);
-    }
-    const chunks: Uint8Array[] = [];
-    for await (const chunk of body as AsyncIterable<Uint8Array>) {
-        chunks.push(chunk);
-    }
-    return Buffer.concat(chunks);
-};
-
-const EMPTY_THEME: DataAppThemeContext = {
-    instructions: null,
-    assets: [],
-    skippedAssetCount: 0,
-};
+import { readS3ObjectAsBuffer } from './s3Utils';
 
 export async function readDesignForDownload(args: {
     s3Client: S3Client;
@@ -48,7 +24,7 @@ export async function readDesignForDownload(args: {
     } = args;
 
     if (!designUuid) {
-        return EMPTY_THEME;
+        return { instructions: null, assets: [], skippedAssetCount: 0 };
     }
 
     const design = await organizationDesignModel.findInOrganization(
@@ -59,7 +35,7 @@ export async function readDesignForDownload(args: {
         logger.warn(
             `Theme ${designUuid} not found during download context build; returning empty theme`,
         );
-        return EMPTY_THEME;
+        return { instructions: null, assets: [], skippedAssetCount: 0 };
     }
 
     const instructionFiles = design.files.filter(
